api/pyproject.toml
[project]
name = "hub-api"
version = "0.1.0"
description = "AI Knowledge Hub API"
requires-python = ">=3.11"
dependencies = [
  "fastapi~=0.115",
  "uvicorn[standard]~=0.30",
  "pydantic~=2.8",
  "SQLAlchemy~=2.0",
  "psycopg[binary]~=3.2",
  "redis~=5.0",
  "python-jose[cryptography]~=3.3",
  "passlib[bcrypt]~=1.7",
  "httpx~=0.27",
  "python-multipart~=0.0.9",
  "tenacity~=9.0",
  "numpy~=1.26",
  "pandas~=2.2",
  "scikit-learn~=1.5",
  "sentence-transformers~=3.0",
  "alembic~=1.13",
  "requests~=2.32"
]

[tool.uvicorn]
factory = true
